moved presseble to homescreen from card, removed unused imports
